+++ src/test/scala/sangria/validation/rules/KnownArgumentNamesSpec.scala
@@ -116,22 +116,5 @@ class KnownArgumentNamesSpec extends WordSpec with ValidationSupport {
         "Unknown argument 'unknown' on field 'doesKnowCommand' of type 'Dog'." → Some(Pos(4, 29)),
         "Unknown argument 'unknown' on field 'doesKnowCommand' of type 'Dog'." → Some(Pos(9, 33))
       ))
-
-    "args may be on object but not interface" in expectFails(
-      """
-        fragment nameSometimesHasArg on Being {
-          name(surname: true)
-          ... on Human {
-            name(surname: true)
-          }
-          ... on Dog {
-            name(surname: true)
-          }
-        }
-      """,
-      List(
-        "Unknown argument 'surname' on field 'name' of type 'Being'." → Some(Pos(3, 16)),
-        "Unknown argument 'surname' on field 'name' of type 'Dog'." → Some(Pos(8, 18))
-      ))
   }
 }